arch: common: timing: Fix timing cycles 32bit rollover
Fix arch_timing_cycles_get() to prevent overflow on 32bit cycles rollover.

The issue is observable, for example when `tests/benchmarks/wait_queues`
or `tests/benchmarks/sched_queues` are executed on qemu for `mps2/an385`
and the benchmark has its iterations large enough as the default
BENCHMARK_NUM_ITERATIONS=1000.

+++ arch/common/timing.c
@@ -28,7 +28,7 @@ timing_t arch_timing_counter_get(void)
 uint64_t arch_timing_cycles_get(volatile timing_t *const start,
 				volatile timing_t *const end)
 {
-	return (*end - *start);
+	return ((uint32_t)*end - (uint32_t)*start);
 }
